requirements.txt
# PostgreSQL
psycopg2-binary>=2.9.9

# Redis
redis>=5.0.1

# MongoDB
pymongo>=4.6.1

# Qdrant
qdrant-client>=1.7.0

# Utilities
numpy>=1.24.3

# LLM & AI
langchain>=0.1.0
langchain-community>=0.0.20
ollama>=0.1.6
sentence-transformers>=2.3.1
chromadb>=0.4.22

# Web UI
streamlit>=1.31.0
streamlit-chat>=0.1.1
